requirements.txt
cbor2==5.7.1
certifi==2025.10.5
charset-normalizer==3.4.4
colorzero==2.0
gpiozero==2.0.1
gtfs-realtime-bindings==1.0.0
idna==3.11
lgpio==0.2.2.0
luma.core==2.5.1
luma.oled==3.14.0
pillow==12.0.0
protobuf==6.33.0
requests==2.32.5
setuptools==80.9.0
smbus2==0.5.0
urllib3==2.5.0
